Rewrite and simplify crate
