+++ Cargo.toml
@@ -8,13 +8,11 @@ repository = "https://github.com/rust-osdev/bootimage"
 edition = "2018"
 
 [dependencies]
-rayon = "1.0"
-toml = "0.5.0"
-wait-timeout = "0.2"
+toml = "0.5.6"
+wait-timeout = "0.2.0"
 llvm-tools = "0.1.1"
-locate-cargo-manifest = "0.1.0"
-json = "0.11.13"
-
-[dependencies.cargo_metadata]
-version = "0.7.4"
-default-features = false
+locate-cargo-manifest = "0.2.0"
+json = "0.12.4"
+anyhow = "1.0.28"
+thiserror = "1.0.16"
+cargo_metadata = "0.9.1"
