--- src/help/build_help.txt
@@ -1,12 +1,0 @@
-Creates a bootable disk image from a Rust kernel
-
-USAGE:
-    bootimage build [BUILD_OPTS]        Create a bootable disk image
-
-    (for other forms of usage see `bootimage --help`)
-
-BUILD_OPTS:
-    Any options are directly passed to `cargo build` (see
-    `cargo build --help` for possible options). After building, a bootloader
-    is downloaded and built, and then combined with the kernel into a bootable
-    disk image.
