+++ src/help/help.txt
@@ -1,31 +1,11 @@
 Creates a bootable disk image from a Rust kernel
 
 USAGE:
-    bootimage [OPTIONS]                         Help and version information
-    bootimage build [BUILD_OPTS]                Create a bootable disk image
-    bootimage run [BUILD_OPTS] -- [RUN_OPTS]    Build and run a disk image
-    bootimage test [BUILD_OPTS]                 Runs integration tests
-    bootimage runner EXECUTABLE                 Convert and run an executable
-
     cargo bootimage [BUILD_OPTS]                Create a bootable disk image
-                                                (equivalent to bootimage build)
+    bootimage runner EXECUTABLE [RUN_OPTS]      Convert and run an executable
 
-For more information about a subcommand run `bootimage [subcommand] --help`.
+For more information about a subcommand run `[subcommand] --help`.
 
-OPTIONS:
-    -h, --help      Prints help information and exit
+GENERAL OPTIONS:
+    -h, --help     Prints help information and exit
     --version      Prints version information and exit
-
-BUILD_OPTS:
-    Any options are directly passed to `cargo build` (see
-    `cargo build --help` for possible options). After building, a bootloader
-    is downloaded and built, and then combined with the kernel into a bootable
-    disk image.
-
-    For configuration options see `bootimage build --help`.
-
-RUN_OPTS:
-    Any options are directly passed to the run command. Note that the run
-    options must be separated from the build options by a "--".
-
-    For configuration options see `bootimage run --help`.
