+++ src/help/mod.rs
@@ -1,38 +1,22 @@
-use crate::ErrorMessage;
-
 const HELP: &str = include_str!("help.txt");
-const BUILD_HELP: &str = include_str!("build_help.txt");
 const CARGO_BOOTIMAGE_HELP: &str = include_str!("cargo_bootimage_help.txt");
-const RUN_HELP: &str = include_str!("run_help.txt");
 const RUNNER_HELP: &str = include_str!("runner_help.txt");
-const TEST_HELP: &str = include_str!("test_help.txt");
 
-pub(crate) fn help() {
+/// Prints a general help text.
+pub fn print_help() {
     print!("{}", HELP);
 }
 
-pub(crate) fn build_help() {
-    print!("{}", BUILD_HELP);
-}
-
-pub(crate) fn cargo_bootimage_help() {
+/// Prints the help for the `cargo bootimage` command.
+pub fn print_cargo_bootimage_help() {
     print!("{}", CARGO_BOOTIMAGE_HELP);
 }
-
-pub(crate) fn run_help() {
-    print!("{}", RUN_HELP);
-}
-
-pub(crate) fn runner_help() {
+/// Prints the help for the `bootimage runner` command.
+pub fn print_runner_help() {
     print!("{}", RUNNER_HELP);
 }
 
-pub(crate) fn test_help() {
-    print!("{}", TEST_HELP);
-}
-
-pub(crate) fn no_subcommand() -> ErrorMessage {
-    "Please invoke `bootimage` with a subcommand (e.g. `bootimage build`).\n\n\
-     See `bootimage --help` for more information."
-        .into()
+/// Prints the version of this crate.
+pub fn print_version() {
+    println!("bootimage {}", env!("CARGO_PKG_VERSION"));
 }
