--- src/subcommand.rs
@@ -1,2 +1,0 @@
-pub mod build;
-pub mod runner;
